+++ packages/client-core/src/index.ts
@@ -4,3 +4,4 @@ export * from './common/modelContext'
 export * from './common/types'
 export * from './common/modules'
 export * from './common/ModuleFactory'
+export * from './hooks'
